Move ScreensDisplay into separate file
